refactored: moved StatsEndpointsEnum to StatsServiceLogic
