+++ WordPress/src/main/java/org/wordpress/android/ui/stats/service/StatsService.java
@@ -49,157 +49,6 @@ public class StatsService extends Service implements StatsServiceLogic.ServiceCo
 
     private StatsServiceLogic mStatsServiceLogic;
 
-    public enum StatsEndpointsEnum {
-        VISITS,
-        TOP_POSTS,
-        REFERRERS,
-        CLICKS,
-        GEO_VIEWS,
-        AUTHORS,
-        VIDEO_PLAYS,
-        COMMENTS,
-        FOLLOWERS_WPCOM,
-        FOLLOWERS_EMAIL,
-        COMMENT_FOLLOWERS,
-        TAGS_AND_CATEGORIES,
-        PUBLICIZE,
-        SEARCH_TERMS,
-        INSIGHTS_POPULAR,
-        INSIGHTS_ALL_TIME,
-        INSIGHTS_TODAY,
-        INSIGHTS_LATEST_POST_SUMMARY,
-        INSIGHTS_LATEST_POST_VIEWS;
-
-        public String getRestEndpointPath() {
-            switch (this) {
-                case VISITS:
-                    return "visits";
-                case TOP_POSTS:
-                    return "top-posts";
-                case REFERRERS:
-                    return "referrers";
-                case CLICKS:
-                    return "clicks";
-                case GEO_VIEWS:
-                    return "country-views";
-                case AUTHORS:
-                    return "top-authors";
-                case VIDEO_PLAYS:
-                    return "video-plays";
-                case COMMENTS:
-                    return "comments";
-                case FOLLOWERS_WPCOM:
-                    return "followers?type=wpcom";
-                case FOLLOWERS_EMAIL:
-                    return "followers?type=email";
-                case COMMENT_FOLLOWERS:
-                    return "comment-followers";
-                case TAGS_AND_CATEGORIES:
-                    return "tags";
-                case PUBLICIZE:
-                    return "publicize";
-                case SEARCH_TERMS:
-                    return "search-terms";
-                case INSIGHTS_POPULAR:
-                    return "insights";
-                case INSIGHTS_ALL_TIME:
-                    return "";
-                case INSIGHTS_TODAY:
-                    return "summary";
-                case INSIGHTS_LATEST_POST_SUMMARY:
-                    return "posts";
-                case INSIGHTS_LATEST_POST_VIEWS:
-                    return "post";
-                default:
-                    AppLog.i(T.STATS, "Called an update of Stats of unknown section!?? " + this.name());
-                    return "";
-            }
-        }
-
-        public StatsEvents.SectionUpdatedAbstract getEndpointUpdateEvent(final long siteId,
-                                                                         final StatsTimeframe timeframe,
-                                                                         final String date,
-                                                                         final int maxResultsRequested,
-                                                                         final int pageRequested,
-                                                                         final BaseStatsModel data) {
-            switch (this) {
-                case VISITS:
-                    return new StatsEvents.VisitorsAndViewsUpdated(siteId, timeframe, date,
-                            maxResultsRequested, pageRequested,
-                            (VisitsModel) data);
-                case TOP_POSTS:
-                    return new StatsEvents.TopPostsUpdated(siteId, timeframe, date,
-                            maxResultsRequested, pageRequested,
-                            (TopPostsAndPagesModel) data);
-                case REFERRERS:
-                    return new StatsEvents.ReferrersUpdated(siteId, timeframe, date,
-                            maxResultsRequested, pageRequested, (ReferrersModel) data);
-                case CLICKS:
-                    return new StatsEvents.ClicksUpdated(siteId, timeframe, date,
-                            maxResultsRequested, pageRequested, (ClicksModel) data);
-                case AUTHORS:
-                    return new StatsEvents.AuthorsUpdated(siteId, timeframe, date,
-                            maxResultsRequested, pageRequested, (AuthorsModel) data);
-                case GEO_VIEWS:
-                    return new StatsEvents.CountriesUpdated(siteId, timeframe, date,
-                            maxResultsRequested, pageRequested, (GeoviewsModel) data);
-                case VIDEO_PLAYS:
-                    return new StatsEvents.VideoPlaysUpdated(siteId, timeframe, date,
-                            maxResultsRequested, pageRequested,
-                            (VideoPlaysModel) data);
-                case SEARCH_TERMS:
-                    return new StatsEvents.SearchTermsUpdated(siteId, timeframe, date,
-                            maxResultsRequested, pageRequested,
-                            (SearchTermsModel) data);
-                case COMMENTS:
-                    return new StatsEvents.CommentsUpdated(siteId, timeframe, date,
-                            maxResultsRequested, pageRequested, (CommentsModel) data);
-                case COMMENT_FOLLOWERS:
-                    return new StatsEvents.CommentFollowersUpdated(siteId, timeframe, date,
-                            maxResultsRequested, pageRequested,
-                            (CommentFollowersModel) data);
-                case TAGS_AND_CATEGORIES:
-                    return new StatsEvents.TagsUpdated(siteId, timeframe, date,
-                            maxResultsRequested, pageRequested, (TagsContainerModel) data);
-                case PUBLICIZE:
-                    return new StatsEvents.PublicizeUpdated(siteId, timeframe, date,
-                            maxResultsRequested, pageRequested, (PublicizeModel) data);
-                case FOLLOWERS_WPCOM:
-                    return new StatsEvents.FollowersWPCOMUdated(siteId, timeframe, date,
-                            maxResultsRequested, pageRequested,
-                            (FollowersModel) data);
-                case FOLLOWERS_EMAIL:
-                    return new StatsEvents.FollowersEmailUdated(siteId, timeframe, date,
-                            maxResultsRequested, pageRequested,
-                            (FollowersModel) data);
-                case INSIGHTS_POPULAR:
-                    return new StatsEvents.InsightsPopularUpdated(siteId, timeframe, date,
-                            maxResultsRequested, pageRequested,
-                            (InsightsPopularModel) data);
-                case INSIGHTS_ALL_TIME:
-                    return new StatsEvents.InsightsAllTimeUpdated(siteId, timeframe, date,
-                            maxResultsRequested, pageRequested,
-                            (InsightsAllTimeModel) data);
-                case INSIGHTS_TODAY:
-                    return new StatsEvents.VisitorsAndViewsUpdated(siteId, timeframe, date,
-                            maxResultsRequested, pageRequested,
-                            (VisitsModel) data);
-                case INSIGHTS_LATEST_POST_SUMMARY:
-                    return new StatsEvents.InsightsLatestPostSummaryUpdated(siteId, timeframe, date,
-                            maxResultsRequested, pageRequested,
-                            (InsightsLatestPostModel) data);
-                case INSIGHTS_LATEST_POST_VIEWS:
-                    return new StatsEvents.InsightsLatestPostDetailsUpdated(siteId, timeframe, date,
-                            maxResultsRequested, pageRequested,
-                            (InsightsLatestPostDetailsModel) data);
-                default:
-                    AppLog.e(T.STATS, "Can't find an Update Event that match the current endpoint: " + this.name());
-            }
-
-            return null;
-        }
-    }
-
     @Override
     public void onCreate() {
         super.onCreate();
